pyproject.toml
[project]
name = "knowcode"
version = "0.2.1"
description = "Efficient codebase knowledge graph builder"
readme = "README.md"
requires-python = ">=3.9, <3.13"
dependencies = [
    "click>=8.1",
    "networkx>=3.0",
    "pyyaml>=6.0",
    "pathspec>=0.11",
    "tree-sitter==0.21.3",
    "tree-sitter-languages>=1.10.0",
    "GitPython>=3.1.0",
    "tiktoken>=0.7.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "openai>=1.0.0",
    "faiss-cpu>=1.7.0",
    "numpy>=1.24.0",
    "watchdog>=3.0.0",
    "google-genai>=0.3.0",
    "google-api-core>=2.29.0",
]

[project.optional-dependencies]
mcp = ["mcp>=1.0.0"]
voyageai = ["voyageai>=0.2.0"]

[project.scripts]
knowcode = "knowcode.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/knowcode"]

[dependency-groups]
dev = [
    "pytest>=7.4",
    "ruff>=0.1.5",
    "mypy>=1.7",
    "types-pyyaml",
    "pytest-cov",
    "mkdocs",
    "mkdocs-material",
    "mkdocstrings[python]",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.mypy]
strict = true
ignore_missing_imports = true
